Refactor existing IssuePrintFormatter logic into RichIssuePrinter
By pulling the decision logic for what will be printed up to the
NewCmdView function, we were able to tease the functionality previously in
the IssuePrintFormatter out and into two new structs. This completes the
refactoring of the previous "human" readable printing into the
RichIssuePrinter and separates the different data structures into their
own files.

Additionally, I've changed the PresentationIssue being passed around from
a pointer to a copy of the object to avoid any unintended side-effects in
the various methods
